+++ R/do_DimPlot.R
@@ -1022,7 +1022,21 @@ do_DimPlot <- function(sample,
                         panel.background = ggplot2::element_rect(fill = "white", color = "white"),
                         legend.background = ggplot2::element_rect(fill = "white", color = "white"))
   }
-
+  
+  # Add font.family to geom_text and geom_label.
+  if (length(p$patches$plots) > 0){
+    num.plots <- length(p$patches$plots)
+  } else {
+    num.plots <- 1
+  }
+  for (plot.use in seq(1, num.plots)){
+    for (layer.use in seq(1, length(p[[plot.use]]$layers))){
+      if (sum(stringr::str_detect(class(p[[plot.use]]$layers[[layer.use]]$geom), "GeomText|GeomLabel"))){
+        p[[plot.use]]$layers[[layer.use]]$aes_params$family <- font.family
+      }
+    }
+  }
+  
 
   # Add plot title to the plots.
   if (!is.null(plot.title)){